Moved assert from ygm::detail to ygm::utility.
